fixed some stuff
