+++ src/main/java/com/app/controllers/CompetitionController.java
@@ -1,14 +1,11 @@
 package com.app.controllers;
 
 import com.app.models.CompetitionModel;
-import com.app.models.DisciplineModel;
 import com.app.models.MemberModel;
 import com.app.models.ResultModel;
 import com.app.models.exceptions.MemberNotFoundException;
 import com.app.models.services.CompetitionService;
 import com.app.models.services.ConfigService;
-import com.app.models.types.GenderType;
-import com.app.models.types.StyleType;
 import com.app.views.CompetitionView;
 
 import java.io.IOException;
@@ -73,14 +70,17 @@ public class CompetitionController {
       VIEW.printInline("Competition ID: ");
       CompetitionModel competition = InputController.validateCompetitionsId(competitions);
 
-      MEMBER_CONTROLLER.viewTableMembers();
-      VIEW.printInline("Member ID: ");
-      MemberModel member =
-          getMember(InputController.validateMemberId(new MemberController().getMembers()));
       do {
-        addResultToCompetition(competition, resultController.addResultTime(member,competition));
-        VIEW.printInline("Do you wish to add another result to this member [Y/n]: ");
-      } while (InputController.promptYesNo());
+        MEMBER_CONTROLLER.viewTableMembers();
+        VIEW.printInline("Member ID: ");
+        MemberModel member =
+            getMember(InputController.validateMemberId(new MemberController().getMembers()));
+        do {
+          addResultToCompetition(competition, resultController.addResultTime(member, competition));
+          VIEW.printInline("Do you wish to add another result to this member [Y/N]: ");
+        } while (InputController.promptYesNo());
+        VIEW.printInline("Do you wish to add results for another member, on this competition [Y/N]: ");
+      }while (InputController.promptYesNo());
     }
   }
 
@@ -147,36 +147,6 @@ public class CompetitionController {
   public void addResultToCompetition(CompetitionModel competition, ResultModel resultModel) {
     competition.addResult(resultModel);
     saveCompetitions();
-  }
-  /** @return a String Array of converted styles */
-  public String[] styleToArray() {
-    String[] result = new String[StyleType.values().length];
-
-    for (int i = 0; i < result.length; i++) {
-      result[i] = StyleType.values()[i].name();
-    }
-
-    return result;
-  }
-
-  /**
-   * Creates an array of the distances available based on swim style and gender
-   *
-   * @param style the style we wish to filter on
-   * @param gender the gender we wish to filter on
-   * @return a String array with filtered distances
-   */
-  public String[] distanceToArray(StyleType style, GenderType gender) {
-    DisciplinesController disciplinesController = new DisciplinesController();
-    ArrayList<DisciplineModel> disciplineModels =
-        disciplinesController.chosenDiscipline(gender, style);
-    String[] result = new String[disciplineModels.size()];
-
-    for (int i = 0; i < result.length; i++) {
-      result[i] = String.valueOf(disciplineModels.get(i).getDistance());
-    }
-
-    return result;
   }
 
   /** Saves the competitions to a file */
