+++ src/main/java/com/app/controllers/ResultController.java
@@ -5,18 +5,16 @@ import com.app.models.DisciplineModel;
 import com.app.models.MemberModel;
 import com.app.models.ResultModel;
 import com.app.models.types.DistanceType;
-import com.app.models.types.GenderType;
 import com.app.models.types.StyleType;
 import com.app.views.ResultView;
 
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 
 public class ResultController {
 
   private final ResultView VIEW = new ResultView();
-  private final CompetitionController COMPETITION_CONTROLLER = new CompetitionController();
+  private final DisciplinesController DISC_CONTROLLER = new DisciplinesController();
 
 
   /**
@@ -26,10 +24,10 @@ public class ResultController {
    * @param competition The competition that we want to add a result to
    */
   public ResultModel addResultTime(MemberModel member, CompetitionModel competition) {
-    VIEW.displayOptions(styleToArray());
-    int styleChoice = InputController.validateOptionRange(styleToArray().length);
+    VIEW.displayOptions(DISC_CONTROLLER.styleToArray());
+    int styleChoice = InputController.validateOptionRange(DISC_CONTROLLER.styleToArray().length);
 
-    String[] distances = distanceToArray(StyleType.values()[styleChoice - 1], member.getGender());
+    String[] distances = DISC_CONTROLLER.distanceToArray(StyleType.values()[styleChoice - 1], member.getGender());
     VIEW.displayOptions(distances);
     int distanceChoice = InputController.validateOptionRange(distances.length);
 
@@ -47,36 +45,4 @@ public class ResultController {
             DistanceType.values()[distanceChoice - 1], StyleType.values()[styleChoice - 1]);
     return  new ResultModel(member, time, disciplineModel, competition, placement);
   }
-
-  /** @return a String Array of converted styles */
-  public String[] styleToArray() {
-    String[] result = new String[StyleType.values().length];
-
-    for (int i = 0; i < result.length; i++) {
-      result[i] = StyleType.values()[i].name();
-    }
-
-    return result;
-  }
-
-  /**
-   * Creates an array of the distances available based on swim style and gender
-   *
-   * @param style the style we wish to filter on
-   * @param gender the gender we wish to filter on
-   * @return a String array with filtered distances
-   */
-  public String[] distanceToArray(StyleType style, GenderType gender) {
-    DisciplinesController disciplinesController = new DisciplinesController();
-    ArrayList<DisciplineModel> disciplineModels =
-        disciplinesController.chosenDiscipline(gender, style);
-    String[] result = new String[disciplineModels.size()];
-
-    for (int i = 0; i < result.length; i++) {
-      result[i] = String.valueOf(disciplineModels.get(i).getDistance());
-    }
-
-    return result;
-  }
-
 }
